fix(mode): extract defaultViewModeForViewport + fix stale-singleton in dispatcher test + correct BookmarksPanel comment
Aristotle iter-1 follow-up on #401.

1. Layer violation: `defaultViewModeForViewport` lived in `tabKind.ts`,
   dragging `viewportStore` (and its matchMedia listener initializer) into
   the previously-pure classifier module. Every importer of `getExtension`
   / `getTabKind` / `tabSupportsReading` would transitively grab those
   listeners. Extracted to `src/lib/viewport.ts`. `tabSupportsReading`
   stays in `tabKind.ts` (pure predicate). The two callers (openFileAsTab
   markdown branch + EditorPane wiki-link handler) updated to import from
   `lib/viewport`. `tabKind.test.ts` no longer carries the helper's
   describe block; it lives in the new `viewport.test.ts` next to the
   source. `EditorPane.wikiLinkMobileMode.test.ts`'s mock target moved
   from `lib/tabKind` to `lib/viewport`.

2. Stale-singleton bug in `openFileAsTab.test.ts`: after `vi.resetModules()`,
   the dynamic `import("../../store/tabStore")` inside each `it` body
   produced a fresh module whose `_core` writable was distinct from the
   top-level `tabStore` import. The `tabStore._reset()` in `beforeEach`
   silently no-op'd against the stale instance, leaking state between
   tests (false-green risk). Moved the reset INSIDE `loadDispatcher`,
   running it AFTER the dynamic import — against the fresh singleton —
   and threaded the resolved `ts` back to each test alongside the dispatch
   function. The `beforeEach` now only resets `readFile` + module graph.

3. Misleading comment in `BookmarksPanel.handleOpenInSplit`. Old comment
   claimed an async/sync race with `moveToPane`, but `openFileAsTab` for
   unencrypted markdown resolves after one microtask — before
   `moveToPane` could fire — so there is no actual race. The real
   constraint is that `handleOpenInSplit` is a synchronous event handler
   and cannot `await` without restructuring the call site. Comment
   rewritten to state that.

Verified:
- `pnpm typecheck` clean.
- `pnpm test` — 148 files / 1404 tests green (+1 file vs prior: `viewport.test.ts`).
- `pnpm build` clean.
